Fix hostname security defaults...
Realized this morning while doing a project for a client that the hostname security defaults lead to a confusing situation that cannot be fixed unless you modify a line in the main program script.  I have not only fixed that, but, I have also made a reasonable addition to the config file so you know how to use the IP/hostname restriction (such as it currently exists).