Sample Script 1.0.1 Update
Updated sample script to work with 1.0.1 (tested with Ubuntu12 using official package and liquidsoap-plugin-lame add-on).  Now uses a path of files in place of .m3u playlist.  Also removes the fallback MP3 requirement and just uses blank() with 5-second duration.  Still need to adjust the media path and password.  However, it does reduce some of the silliness a bit.